+++ javascript/timers.js
@@ -7,7 +7,7 @@ async function timer(name, elapsed) {
 async function logTimers() {
   allTimers.sort((a, b) => b[1] - a[1]);
   const filteredTimers = allTimers.filter((t) => t[1] > 50);
-  log('timers', filteredTimers);
+  debug('startupTimers', filteredTimers);
   // xhrPost(`${window.api}/log`, { debug: JSON.stringify(filteredTimers) });
 }